Change keep-alive method to refreshing token on interval
An authentication token only allows access to Cloud Files for 24 hours.
After this time, the client must reauthenticate. In the previous Rackit
implementation, authentication was performed once on initialization,
and then subsequently whenever a 401 error was received. This method
complicates streaming uploads, as the source may need to be piped
twice. To resolve this, Rackit has been changed to automatically
reauthenticate every 23 hours, and hopefuly avoid any 401 responses.